basic function passed, the logic of client is wrong, user must be able to regain the online user list from server, gain the list from server on initialization is unnessesary.

+++ TCPChatServer/TCPChatServer/Support.h
@@ -11,6 +11,6 @@
 
 #include "includes.h"
 
-char * StrCut(const char * original, int startIndex, long cutLength);
+char * StrCut(const char * original, long startIndex, long cutLength);
 
 #endif /* Support_h */
